exceptions are added, bugs fixed and frontend improved

+++ src/main/java/com/library/libraryProject/service/impl/BookServiceImpl.java
@@ -1,42 +1,62 @@
 package com.library.libraryProject.service.impl;
 
+import com.library.libraryProject.exception.BadResourceException;
+import com.library.libraryProject.exception.ResourceAlreadyExistsException;
 import com.library.libraryProject.model.Book;
 import com.library.libraryProject.repository.BookRepository;
 import com.library.libraryProject.service.BookService;
 import lombok.RequiredArgsConstructor;
-import org.modelmapper.ModelMapper;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.data.rest.webmvc.ResourceNotFoundException;
 import org.springframework.stereotype.Service;
 
-import java.util.Arrays;
 import java.util.List;
 
 @Service
 @RequiredArgsConstructor
 public class BookServiceImpl implements BookService {
-    private final ModelMapper modelMapper;
     private final BookRepository bookRepository;
 
-    public List<Book> getAll() {
-        List<Book> data = bookRepository.findAll();
-        return Arrays.asList(modelMapper.map(data, Book[].class));
+    private boolean existsById(Long id) {
+        return bookRepository.existsById(id);
     }
 
     @Override
-    public Book getById(Long id) {
-        return bookRepository.findById(id).orElse(null);
+    public List<Book> getAllBooks() {
+        return bookRepository.findAll();
     }
-    //public Book getById(Long id) { return bookRepository.getOne(id); }
 
     @Override
-    public Book save(Book book) {
-        return bookRepository.save(book);
+    public Book getById(Long id) throws ResourceNotFoundException {
+        //return bookRepository.findById(id).orElse(null);
+        Book book = bookRepository.findById(id).orElse(null);
+        if (book==null) {
+            throw new ResourceNotFoundException("Cannot find book with id: " + id);
+        }
+        else return book;
+    }
+
+    @Override
+    public Book save(Book book) throws BadResourceException, ResourceAlreadyExistsException {
+        //return bookRepository.save(book);
+        if (!StringUtils.isEmpty(book.getBook_name())) {
+            if (book.getBook_id() != null && existsById(book.getBook_id())) {
+                throw new ResourceAlreadyExistsException("Book with id: " + book.getBook_id() + " already exists");
+            }
+            return bookRepository.save(book);
+        }
+        else {
+            BadResourceException exc = new BadResourceException("Failed to save contact");
+            exc.addErrorMessage("Book is null or empty");
+            throw exc;
+        }
     }
 
     @Override
     public Book update(Long id, Book book) {
         Book bookUpdate = bookRepository.getOne(id);
         if(bookUpdate==null){
-            throw  new IllegalArgumentException("Book does not exist with this id");
+            throw new ResourceNotFoundException("Cannot find book with id: " + id);
         }
         bookUpdate.setBook_name(book.getBook_name());
         bookUpdate.setIsbn(book.getIsbn());
@@ -54,7 +74,13 @@ public class BookServiceImpl implements BookService {
 */
     @Override
     public Boolean delete(Long id) {
-        bookRepository.deleteById(id);
-        return null;
+        if (!existsById(id)) {
+            throw new ResourceNotFoundException("Cannot find contact with id: " + id);
+        }
+        else {
+            bookRepository.deleteById(id);
+            return null;
+        }
     }
+
 }
